requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas>=1.5.0,<2.0.0  # Compatible with Python 3.8
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Testing dependencies
requests>=2.31.0  # For API testing scripts

# Phase 2 Dependencies (for GenAI integration)
# google-generativeai==0.3.0  # Uncomment when implementing Phase 2

